Fix custom age groups
The expressions were not correct - processing algorithm's input parameters
are not available in expressions automatically neither as fields (columns) nor
as variables. One way of fixing (with minimal changes) would be to use string
concatenation in python and do something like:
  'eval( ... generate_series(' + parameters['Uppersecondarystartingage'] + ', ...)'
But the whole expression is already quite complex, so the fix comes
with a simplification to directly build the expressions in Python without
resorting to complex QGIS expressions.